backend/requirements.txt
streamlit==1.44.1
pandas==2.2.3
pydantic==2.11.3
requests==2.32.3
mysql-connector-python==9.2.0
fastapi==0.115.12
matplotlib==3.10.1
python-dotenv
uvicorn

# Optional but recommended
jinja2==3.1.3
email-validator==2.1.1
python-multipart==0.0.9

# For sending emails (SMTP)
smtplib3==1.0.0.post1